Add hash module
